Fixing fuzzy images on related products and problem with msrp
Related products also use an MSRP hide message which was being moved
into the add to cart wrapper, whoops! Had to create a new template in
order to change the image size for related products, I grabbed 300
pixels because thats the value we use on the catalog category view as
well.